requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
openai==1.3.0
requests==2.31.0
aiogram==3.3.0
asyncpg==0.31.0
sqlalchemy==2.0.44
python-dotenv==1.2.1
